Fix warnings due to camparing different types.

+++ lib/cola.cpp
@@ -41,7 +41,7 @@ void cola_kick(Particles* const particles, const double avel1)
 
   
   Particle* const p= particles->p;
-  const int np= particles->np_local;
+  const size_t np= particles->np_local;
   Float3* const f= particles->force;
 
   
@@ -88,7 +88,7 @@ void cola_drift(Particles* const particles, const double apos1)
 #ifdef _OPENMP
   #pragma omp parallel for default(shared)
 #endif
-  for(int i=0; i<np; i++) {
+  for(size_t i=0; i<np; i++) {
     p[i].x[0] += p[i].v[0]*dt + 
                  (p[i].dx1[0]*da1 + p[i].dx2[0]*da2);
     p[i].x[1] += p[i].v[1]*dt +
